add FirstClassObject protocol to extension Dictionary and Array
