+++ JSONParser/JSONParser/DataInfo.swift
@@ -9,7 +9,6 @@
 import Foundation
 
 struct DataInfo {
-    var type = ""
     var countOfNumber = 0
     var countOfString = 0
     var countOfBool = 0
